+++ packages/components-js/tests/e2e/specs/tabs-bar.e2e.ts
@@ -4,11 +4,11 @@ import {
   getAttribute,
   getBrowser,
   getElementPositions,
-  getProperty,
+  getProperty, getStyleOnFocus,
   initAddEventListener,
   reattachElement,
-  selectNode,
-  setContentWithDesignSystem,
+  selectNode, setAttribute,
+  setContentWithDesignSystem, waitForInheritedCSSTransition, expectedStyleOnFocus,
   waitForStencilLifecycle,
 } from '../helpers';
 
@@ -41,7 +41,9 @@ describe('tabs-bar', () => {
     await setContentWithDesignSystem(page, isWrapped ? `<div style="width: 300px">${content}</div>` : content);
   };
 
+  const getHost = () => selectNode(page, 'p-tabs-bar');
   const getAllButtons = () => page.$$('button');
+  const getAllLinks = () => page.$$('a');
   const getScrollArea = () => selectNode(page, 'p-tabs-bar >>> .p-tabs-bar__scroll-area');
   const getStatusBar = () => selectNode(page, 'p-tabs-bar >>> .p-tabs-bar__status-bar');
   const getGradientNext = () => selectNode(page, 'p-tabs-bar >>> .p-tabs-bar__gradient--next');
@@ -525,6 +527,60 @@ describe('tabs-bar', () => {
 
       await page.evaluate(() => console.error('test error'));
       expect(getErrorsAmount(consoleMessages)).toBe(1);
+    });
+  });
+
+  describe('focus state', () => {
+    it('should show outline of slotted <button> when it is focused', async () => {
+      await setContentWithDesignSystem(
+        page,
+        `<p-tabs-bar>
+        <button>Content1</button>
+        <button>Content2</button>
+        <button>Content3</button>
+      </p-tabs-bar>`
+      );
+
+      const host = await getHost();
+      const buttons = await getAllButtons();
+
+      expect(await getStyleOnFocus(buttons[0])).toBe(expectedStyleOnFocus({color: 'active'}));
+      expect(await getStyleOnFocus(buttons[1])).toBe(expectedStyleOnFocus());
+      expect(await getStyleOnFocus(buttons[2])).toBe(expectedStyleOnFocus());
+
+      await setAttribute(host, 'theme', 'dark');
+      await waitForStencilLifecycle(page);
+      await waitForInheritedCSSTransition(page);
+
+      expect(await getStyleOnFocus(buttons[0])).toBe(expectedStyleOnFocus({color: 'active', theme: 'dark'}));
+      expect(await getStyleOnFocus(buttons[1])).toBe(expectedStyleOnFocus({theme: 'dark'}));
+      expect(await getStyleOnFocus(buttons[2])).toBe(expectedStyleOnFocus({theme: 'dark'}));
+    });
+
+    it('should show outline of slotted <a> when it is focused', async () => {
+      await setContentWithDesignSystem(
+        page,
+        `<p-tabs-bar>
+        <a href="#">Content1</a>
+        <a href="#">Content2</a>
+        <a href="#">Content3</a>
+      </p-tabs-bar>`
+      );
+
+      const host = await getHost();
+      const links = await getAllLinks();
+
+      expect(await getStyleOnFocus(links[0])).toBe(expectedStyleOnFocus({color: 'active'}));
+      expect(await getStyleOnFocus(links[1])).toBe(expectedStyleOnFocus());
+      expect(await getStyleOnFocus(links[2])).toBe(expectedStyleOnFocus());
+
+      await setAttribute(host, 'theme', 'dark');
+      await waitForStencilLifecycle(page);
+      await waitForInheritedCSSTransition(page);
+
+      expect(await getStyleOnFocus(links[0])).toBe(expectedStyleOnFocus({color: 'active', theme: 'dark'}));
+      expect(await getStyleOnFocus(links[1])).toBe(expectedStyleOnFocus({theme: 'dark'}));
+      expect(await getStyleOnFocus(links[2])).toBe(expectedStyleOnFocus({theme: 'dark'}));
     });
   });
 });